+++ Deep-EIoU/yolox/yolox/deepsort_tracker/detection.py
@@ -26,7 +26,7 @@ class Detection(object):
     def __init__(self, tlwh, confidence, feature):
         self.tlwh = np.asarray(tlwh, dtype=float)
         self.confidence = float(confidence)
-        self.feature = np.asarray(feature, dtype=float32)
+        self.feature = np.asarray(feature, dtype=float)
 
     def to_tlbr(self):
         """Convert bounding box to format `(min x, min y, max x, max y)`, i.e.,
